+++ aws-iot-log-server/app/aws_iot_logger.py
