+++ aws-iot-log-server/app/boto3_session.py
