+++ aws-iot-log-server/app/configs.py
