+++ aws-iot-log-server/app/custom_http_handler.py
