+++ aws-iot-log-server/app/greengrass_config.py
