+++ aws-iot-log-server/app/http_server.py
