+++ aws-iot-log-server/app/log_proxy_server.py
