+++ aws-iot-log-server/tests/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+import sys
+import os
+
+
+@pytest.fixture(autouse=True)
+def pythonpath():
+    sys.path.append(
+        os.path.abspath(os.path.dirname(os.path.abspath(__file__)) + "/../app/")
+    )
