+++ aws-iot-log-server/tests/data/greengrass/config.json
