+++ aws-iot-log-server/tests/data/greengrass/config_invalid_thingArn.json
