+++ aws-iot-log-server/tests/test_boto3_session.py
@@ -4,8 +4,8 @@ import os
 class TestBoto3Session:
     def test__refresh_credentials(self, mocker, shared_datadir):
         import requests
-        from aws_gglog.boto3_session import Boto3Session
-        from aws_gglog.greengrass_config import GreengrassConfig
+        from boto3_session import Boto3Session
+        from greengrass_config import GreengrassConfig
 
         resp_mock = mocker.MagicMock()
         resp_mock.text = '{"credentials":{"accessKeyId":"123","secretAccessKey":"abc","sessionToken":"ABC","expiration":"2021-10-01T09:18:06Z"}}'
