--- tests/aws_gglog/test_logger.py
@@ -1,123 +1,0 @@
-import logging
-import os
-import pytest
-
-
-class Test_BaseLogger:
-    def test__gen_log_stream_name(self, shared_datadir):
-        from aws_gglog.logger import _BaseLogger
-
-        name = _BaseLogger._get_stream_name("foo-bar")
-        assert name == "{strftime:%Y/%m/%d}/foo-bar"
-
-    @pytest.mark.parametrize(
-        "keys, do_raise",
-        [
-            (
-                (
-                    "AWS_CREDENTIAL_PROVIDER_ENDPOINT",
-                    "AWS_ROLE_ALIAS",
-                    "AWS_CLOUDWATCH_LOG_GROUP",
-                ),
-                False,
-            ),
-            (
-                (
-                    "AWS_CREDENTIAL_PROVIDER_ENDPOINT",
-                    "AWS_ROLE_ALIAS",
-                    "AWS_CLOUDWATCH_LOG_GROUP",
-                    "AWS_GREENGRASS_CONFIG",
-                    "AWS_CA_CERT_FILE",
-                    "AWS_PRIVATE_KEY_FILE",
-                    "AWS_CERT_FILE",
-                    "AWS_REGION",
-                    "AWS_THING_NAME",
-                ),
-                False,
-            ),
-            (
-                (
-                    "AWS_ROLE_ALIAS",
-                    "AWS_CLOUDWATCH_LOG_GROUP",
-                ),
-                True,
-            ),
-        ],
-    )
-    def test__get_config_from_environment(self, mocker, keys, do_raise):
-        from aws_gglog.logger import _BaseLogger
-
-        for key in keys:
-            mocker.patch.dict(os.environ, {key: f"{key.lower()}"})
-
-        if do_raise:
-            with pytest.raises(Exception):
-                _BaseLogger._get_config_from_environment()
-        else:
-            config = _BaseLogger._get_config_from_environment()
-            assert config == {key: key.lower() for key in keys}
-
-    def test_get_instance(self):
-        from aws_gglog.logger import _BaseLogger
-
-        _BaseLogger._instance = None
-        assert _BaseLogger.get_instance() == _BaseLogger.get_instance()
-
-    def test_singleton(self):
-        from aws_gglog.logger import _BaseLogger
-
-        with pytest.raises(Exception) as e:
-            _BaseLogger._instance = None
-            _BaseLogger.get_instance()
-            _BaseLogger()
-
-        assert str(e.value) == "BaseLogger is singleton"
-
-    def test_create_instance_with_cloudwatch_handler(self, mocker):
-        from aws_gglog.logger import _BaseLogger
-        import watchtower
-
-        mocker.patch("boto3_session.Boto3Session.get_session", return_value="session")
-        mocker.patch(
-            "greengrass_config.GreengrassConfig.parse_config",
-            return_value={
-                "ca_cert": "/foo/bar/ca_cert.pem",
-                "private_key": "/foo/bar/private_key.pem",
-                "cert": "/foo/bar/cert.pem",
-                "region": "ap-northeast-1",
-                "thing_name": "foo-bar",
-            },
-        )
-        watchtower.CloudWatchLogHandler = mocker.MagicMock()
-
-        mocker.patch.dict(os.environ, {"AWS_GREENGRASS_CONFIG": "/foo/bar/config.json"})
-        mocker.patch.dict(os.environ, {"AWS_CREDENTIAL_PROVIDER_ENDPOINT": "foo.bar"})
-        mocker.patch.dict(os.environ, {"AWS_ROLE_ALIAS": "foo-bar"})
-        mocker.patch.dict(os.environ, {"AWS_CLOUDWATCH_LOG_GROUP": "foo/bar"})
-
-        _BaseLogger._instance = None
-        base_logger = _BaseLogger()
-        base_logger.set_handlers(
-            logging.WARNING,
-            "[%(asctime)s][%(levelname)s]-%(filename)s:%(lineno)d,%(message)s",
-        )
-
-        watchtower.CloudWatchLogHandler.assert_called_once_with(
-            boto3_session="session",
-            log_group="foo/bar",
-            stream_name="{strftime:%Y/%m/%d}/foo-bar",
-            create_log_group=True,
-            create_log_stream=True,
-            send_interval=4,
-        )
-
-    def test_create_instance_without_cloudwatch_handler(self, mocker):
-        from aws_gglog.logger import _BaseLogger
-        import watchtower
-
-        watchtower.CloudWatchLogHandler = mocker.MagicMock()
-
-        _BaseLogger._instance = None
-        _BaseLogger()
-
-        watchtower.CloudWatchLogHandler.assert_not_called()
